leave grading dir temporary

+++ src/main/java/org/pesho/grader/SubmissionGrader.java
@@ -37,9 +37,11 @@ public class SubmissionGrader {
 	public double grade() {
 		File sandboxDir = new File(originalSourceFile.getParentFile(), "sandbox_"+originalSourceFile.getName());
 		try {
-			return gradeInternal(sandboxDir);
+			double score = gradeInternal(sandboxDir);
+			if (score > 0) FileUtils.deleteQuietly(sandboxDir);
+			return score;
 		} finally {
-			FileUtils.deleteQuietly(sandboxDir);
+//			FileUtils.deleteQuietly(sandboxDir);
 		}
 	}
